Improved tests for JwtBuilder
