+++ src/JWT/JwtEncoder.cs
@@ -46,6 +46,8 @@ namespace JWT
         {
             if (payload is null)
                 throw new ArgumentNullException(nameof(payload));
+            if (!_algorithm.IsAsymmetric && key is null)
+                throw new ArgumentNullException(nameof(key));
 
             var segments = new List<string>(3);
